% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorr.R
\name{getCorr}
\alias{getCorr}
\title{getCorr Function}
\usage{
getCorr(year)
}
\arguments{
\item{year}{The year of interest}
}
\description{
This function allows you to get the correlation matrix for all numeric variables for a certain year.
}
\examples{
getCorr()
}
\keyword{correlation}
