% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanAge.R
\name{getMeanAge}
\alias{getMeanAge}
\title{getMeanAge Function}
\usage{
getMeanAge(year)
}
\arguments{
\item{year}{The year of interest}
}
\description{
This function allows you to get stastical mean for the players' age for a certain year.
}
\examples{
getMeanAge()
}
\keyword{age}
