% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOldestPlayer.R
\name{getOldestPlayer}
\alias{getOldestPlayer}
\title{getOldestPlayer Function}
\usage{
getOldestPlayer(year)
}
\arguments{
\item{year}{The year of interest}
}
\description{
This function allows you to get the oldest player's name for a certain year. If multiple players share the same age, the one with
}
\examples{
getOldestPlayer()
}
\keyword{oldest}
